feat(MW-Java): Fully generated via Polymorph
On my local, all tests are passing.
The only difference b/w CI & my local is
an updated dafny-java-converison.

However, the generated Dafny was not committed.
There are manual changes to the Types file.

Additionally, the Middleware's Shim
(DynamoDbEncryptionMiddlewareInternal.java)
has a manual correction,
as the Dafny Method to create the Middleware Client
is not named as expected.

Polymorph is at c410458